multi_container_demo/predictor/requirements.txt
kafka-python>=1.3.5
tweet-preprocessor==0.6.0